pyproject.toml
[tool.poetry]
name = "home_assistant_nature_remo"
version = "0.1.0"
description = "A Nature Remo integration for Home Assistant"
authors = ["mrtolkien <gary.mialaret@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.1"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
pytest-homeassistant-custom-component = "^0.4.5"
# TODO Remove those once we have our own async package
voluptuous = "^0.12.2"
marshmallow = "^3.7.1"
click = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
